fix: add dependencies for tgbot 18.2.1, smallrye-graphql, and reactive components
This commit adds dependencies for tgbot 18.2.1, smallrye-graphql, and reactive components to the project.

The tgbot dependency has been updated to version 18.2.1 to take advantage of the new features and bug fixes in that release.

The smallrye-graphql dependency has been added to enable the use of GraphQL in the project.

The reactive components dependencies (quarkus-hibernate-reactive-panache-kotlin, quarkus-reactive-pg-client, and quarkus-redis-client) have been added to enable the use of reactive programming in the project.